Adding request validations and DB coalesce

+++ src/main/java/com/example/library/management/tool/library/service/LanguageService.java
@@ -1,9 +1,9 @@
 package com.example.library.management.tool.library.service;
 
 import com.example.library.management.tool.library.dao.LanguageDao;
-import com.example.library.management.tool.library.dto.genre.Genre;
 import com.example.library.management.tool.library.dto.language.Language;
 import com.example.library.management.tool.library.dto.standardresponse.ApiResponse;
+import com.example.library.management.tool.library.util.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,14 +20,23 @@ public class LanguageService {
     }
 
     public ApiResponse addLanguage(String languageName) {
+        if (Validator.isEmptyOrNull(languageName)) {
+            return new ApiResponse(false, "Language name shouldn't be empty or null.");
+        }
         return languageDao.addLanguage(languageName);
     }
 
     public ApiResponse updateLanguage(Language language) {
+        if (Validator.isEmptyOrNull(language.getLanguageName())) {
+            return new ApiResponse(false, "Genre name shouldn't be empty or null.");
+        }
         return languageDao.updateLanguage(language);
     }
 
     public ApiResponse deleteLanguage(int languageId){
+        if (languageId == 0) {
+            return new ApiResponse(false, "Admin id shouldn't or null.");
+        }
         return languageDao.deleteLanguage(languageId);
     }
 }
