+++ src/main/java/com/example/library/management/tool/library/service/UserService.java
@@ -1,9 +1,9 @@
 package com.example.library.management.tool.library.service;
 
 import com.example.library.management.tool.library.dao.UserDao;
-import com.example.library.management.tool.library.dto.genre.Genre;
 import com.example.library.management.tool.library.dto.standardresponse.ApiResponse;
 import com.example.library.management.tool.library.dto.user.User;
+import com.example.library.management.tool.library.util.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,14 +20,29 @@ public class UserService {
     }
 
     public ApiResponse addUser(User user) {
+
+        if (Validator.isEmptyOrNull(user.getUserName()) ||
+        Validator.isEmptyOrNull(user.getUserPhone()) ||
+        Validator.isEmptyOrNull(user.getUserEmail())) {
+            return new ApiResponse(false, "User Name, User Phone and User Email shouldn't be empty or null.");
+        }
+
         return userDao.addUser(user);
     }
 
     public ApiResponse updateUser(User user) {
+        if (Validator.isEmptyOrNull(user.getUserName()) &&
+                Validator.isEmptyOrNull(user.getUserEmail()) &&
+                Validator.isEmptyOrNull(user.getUserPhone())) {
+            return new ApiResponse(false, "User Name, User phone, and User email (all) shouldn't be empty or null.");
+        }
         return userDao.updateUser(user);
     }
 
-    public ApiResponse deleteUser(int userId){
+    public ApiResponse deleteUser(int userId) {
+        if (userId == 0) {
+            return new ApiResponse(false, "User id shouldn't be null.");
+        }
         return userDao.deleteUser(userId);
     }
 }
